:truck: Move file that handle errors to its own package
