+++ src/main/java/com/mercadolivre/apimlv2/shared/errors/ValidationErrorHandler.java
@@ -1,4 +1,4 @@
-package com.mercadolivre.apimlv2.shared;
+package com.mercadolivre.apimlv2.shared.errors;
 
 import org.springframework.context.MessageSource;
 import org.springframework.context.i18n.LocaleContextHolder;
